Use 'spec.ip' instead of looking up the default network
We want to use Bosh CLI v2 but it fails evaluating the
`spec.networks.send(spec.networks.methods(false).first).ip` expression. We
found an old issue about it [1], but they fixed it originally in Bosh v257.1.
In the latest version it doesn't seem to work, but in v258 they introduced
`spec.ip` which shoud return with the same value.

[1] https://github.com/cloudfoundry/bosh/issues/1318